Add command handling
Per-file command handling is added. Command registration is done during
the bot's initialization phase instead of a separate script.

The debug module is also added as a dependency and will act as a verbose
logger. To enable logging, just set the DEBUG environment variable to
' xingqiu:<filename> ' to only enable logging for a single file or
' xingqiu:* ' to enable logging for the whole project. Refer to debug
module's documentation in npm for more details.