Add event handling
Event handling is also done per-file though this is added mainly as
future-proofing the bot, since right now I can't think of ways that
the bot needs that much event handlers to warrant a per-file system
like this.

Another small thing that was added is that if the bot is registering
commands in the test guild, it will also delete all global commands it
has registered. This is in assumption that bots running with a test
guild are running in "test mode." It also prevents commands doubling
when both commands are registered globally and in guild.